Transform current useStyles hook into utils
